[Fix #211] r0NightStartHandler 함수로 별도 분리
