+++ types/index.ts
@@ -2,6 +2,7 @@ import { TrackReferenceOrPlaceholder } from "@livekit/components-react";
 import { User } from "@supabase/supabase-js";
 import { Participant } from "livekit-client";
 import { StaticImageData } from "next/image";
+import { Dispatch, SetStateAction } from "react";
 
 export interface MafiaRoom {
   room: string;
@@ -143,4 +144,21 @@ export interface ShowModalState {
   setMessage: (newMessage: string) => void;
   setTimer: (newTimer: number) => void;
   setIsClose: (newIsClose: boolean) => void;
+}
+
+export interface TimerState {
+  timerIds: NodeJS.Timeout[];
+  setTimerIds: (newTimerId: NodeJS.Timeout) => void;
+}
+
+export interface showModalComponents {
+  userId: string;
+  roomId: string;
+  setIsOpen: (newIsOpen: boolean) => void;
+  setTitle: (newTitle: string) => void;
+  setMessage: (newMessage: string) => void;
+  setTimer: (newTimer: number) => void;
+  setIsClose: (newIsClose: boolean) => void;
+  setIsOverlay: (newIsOverlay: boolean) => void;
+  setTimerIds: Dispatch<SetStateAction<NodeJS.Timeout[]>>;
 }
